vm-files/caption-service/requirements.txt
flask
transformers
torch
Pillow